% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxBoost.R
\name{print.CoxBoost}
\alias{print.CoxBoost}
\title{Print a CoxBoost type}
\usage{
\method{print}{CoxBoost}(x, ...)
}
\arguments{
\item{x}{a CoxBoost object}

\item{...}{not used}
}
\description{
Print a CoxBoost type
}
