% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxBoost.R
\name{summary.CoxBoost}
\alias{summary.CoxBoost}
\title{Summary of a CoxBoost type}
\usage{
\method{summary}{CoxBoost}(object, ...)
}
\arguments{
\item{object}{a CoxBoost object}

\item{...}{not used}
}
\description{
Summary of a CoxBoost type
}
